fix: add thousand separators to currency displays and adjust cooldown
Updated price formatting across commands and crew interactions to include
thousands separators (e.g., $1,234.56) for better readability. Reduced
`/fish` cooldown from 10 to 1 use per 30 seconds to improve user
experience. Fixed parameter type hint in `calculate_portfolio` to `float`
for consistency with monetary values.

+++ code/engines/economy_engine.py
@@ -11,7 +11,7 @@ def get_net_worth_status(total_net_worth: int) -> str:
     else:
         return "Hobbyist Fisherman"
 
-def calculate_portfolio(wallet_cash: int, user_inv: list[tuple[str, int]], market_prices_dict: dict) -> dict:
+def calculate_portfolio(wallet_cash: float, user_inv: list[tuple[str, int]], market_prices_dict: dict) -> dict:
     """Calculates the total value of the player's wallet and assets."""
     inventory_value = 0
